+++ data/src/main/java/com/example/data/storage/DayStorage.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface DayStorage {
     val allDays : Flow<List<DayRepositoryEntity>>
-    suspend fun getDay(date: Int): DayRepositoryEntity?
+    suspend fun getDay(date: Long): DayRepositoryEntity?
 
     suspend fun addDay(day: DayRepositoryEntity) : Boolean
 }
