+++ domain/src/main/java/com/example/domain/repositoriy/DayRepository.kt
@@ -3,6 +3,6 @@ package com.example.domain.repositoriy
 import com.example.domain.models.Day
 
 interface DayRepository {
-    suspend fun getDay(date: Int): Day?
+    suspend fun getDay(date: Long): Day?
     suspend fun addDay(day: Day) : Boolean
 }
