+++ domain/src/main/java/com/example/domain/usecase/GetDayUseCase.kt
@@ -4,7 +4,7 @@ import com.example.domain.models.Day
 import com.example.domain.repositoriy.DayRepository
 
 class GetDayUseCase (private val dayRepository: DayRepository){
-    suspend fun execute(date: Int) : Day? {
+    suspend fun execute(date: Long) : Day? {
         return dayRepository.getDay(date)
     }
 }
